requirements.txt
Django>=4.0
djangorestframework
channels
channels_redis
psycopg2-binary
requests
